+++ com.unity.render-pipelines.universal/Editor/ForwardRendererDataEditor.cs
@@ -11,7 +11,6 @@ namespace UnityEditor.Rendering.Universal
         private static class Styles
         {
             public static readonly GUIContent RendererTitle = new GUIContent("Forward Renderer", "Custom Forward Renderer for Universal RP.");
-            public static readonly GUIContent PostProcessIncluded = EditorGUIUtility.TrTextContent("Post Processing", "Includes or removes Post Processing from builds, this includes rendering passes, shaders and texture resources.");
             public static readonly GUIContent PostProcessLabel = new GUIContent("Post Process Data", "The asset containing references to shaders and Textures that the Renderer uses for post-processing.");
             public static readonly GUIContent FilteringLabel = new GUIContent("Filtering", "Controls filter rendering settings for this renderer.");
             public static readonly GUIContent OpaqueMask = new GUIContent("Opaque Layer Mask", "Controls which opaque layers this renderer draws.");
@@ -58,11 +57,16 @@ namespace UnityEditor.Rendering.Universal
             serializedObject.Update();
 
             EditorGUILayout.Space();
-            EditorGUILayout.LabelField(Styles.RendererTitle, EditorStyles.boldLabel); // Title
-            EditorGUI.indentLevel++;
-            DrawPostProcessArea();
-            EditorGUI.indentLevel--;
-            EditorGUILayout.Space();
+
+            if (m_PostProcessData.objectReferenceValue != null)
+            {
+                EditorGUILayout.LabelField(Styles.RendererTitle, EditorStyles.boldLabel); // Title
+                EditorGUI.indentLevel++;
+                EditorGUILayout.PropertyField(m_PostProcessData, Styles.PostProcessLabel);
+                EditorGUILayout.HelpBox("Post Process Data was moved to UniversalRenderPipelineAsset.", MessageType.Warning);
+                EditorGUI.indentLevel--;
+                EditorGUILayout.Space();
+            }
 
             EditorGUILayout.LabelField(Styles.FilteringLabel, EditorStyles.boldLabel);
             EditorGUI.indentLevel++;
@@ -128,22 +132,6 @@ namespace UnityEditor.Rendering.Universal
                     resources.shaders = null;
                     ResourceReloader.ReloadAllNullIn(target, UniversalRenderPipelineAsset.packagePath);
                 }
-            }
-        }
-
-        private void DrawPostProcessArea()
-        {
-            EditorGUI.BeginChangeCheck();
-            var postProcessIncluded = EditorGUILayout.Toggle(Styles.PostProcessIncluded, m_PostProcessData.objectReferenceValue != null);
-            if (EditorGUI.EndChangeCheck())
-            {
-                m_PostProcessData.objectReferenceValue = postProcessIncluded ? PostProcessData.GetDefaultPostProcessData() : null;
-            }
-            if (postProcessIncluded)
-            {
-                EditorGUI.indentLevel++;
-                EditorGUILayout.PropertyField(m_PostProcessData, Styles.PostProcessLabel);
-                EditorGUI.indentLevel--;
             }
         }
     }