+++ com.unity.render-pipelines.universal/Editor/UniversalRenderPipelineAssetEditor.cs
@@ -56,6 +56,8 @@ namespace UnityEditor.Rendering.Universal
             public static GUIContent supportsSoftShadows = EditorGUIUtility.TrTextContent("Soft Shadows", "If enabled pipeline will perform shadow filtering. Otherwise all lights that cast shadows will fallback to perform a single shadow sample.");
 
             // Post-processing
+            public static GUIContent postProcessLabel = new GUIContent("Post Process Data", "The asset containing references to shaders and Textures that the Renderer uses for post-processing.");
+            public static GUIContent postProcessIncluded = EditorGUIUtility.TrTextContent("Post Processing", "Includes or removes Post Processing from builds, this includes rendering passes, shaders and texture resources.");
             public static GUIContent colorGradingMode = EditorGUIUtility.TrTextContent("Grading Mode", "Defines how color grading will be applied. Operators will react differently depending on the mode.");
             public static GUIContent colorGradingLutSize = EditorGUIUtility.TrTextContent("LUT size", "Sets the size of the internal and external color grading lookup textures (LUTs).");
             public static string colorGradingModeWarning = "HDR rendering is required to use the high dynamic range color grading mode. The low dynamic range will be used instead.";
@@ -137,6 +139,7 @@ namespace UnityEditor.Rendering.Universal
         SerializedProperty m_ShaderVariantLogLevel;
 
         LightRenderingMode selectedLightRenderingMode;
+        SerializedProperty m_PostProcessData;
         SerializedProperty m_ColorGradingMode;
         SerializedProperty m_ColorGradingLutSize;
 
@@ -211,6 +214,7 @@ namespace UnityEditor.Rendering.Universal
 
             m_ShaderVariantLogLevel = serializedObject.FindProperty("m_ShaderVariantLogLevel");
 
+            m_PostProcessData = serializedObject.FindProperty("m_PostProcessData");
             m_ColorGradingMode = serializedObject.FindProperty("m_ColorGradingMode");
             m_ColorGradingLutSize = serializedObject.FindProperty("m_ColorGradingLutSize");
 
@@ -384,6 +388,45 @@ namespace UnityEditor.Rendering.Universal
 
                 EditorGUI.indentLevel++;
 
+                EditorGUI.BeginChangeCheck();
+                var postProcessIncluded = EditorGUILayout.Toggle(Styles.postProcessIncluded, m_PostProcessData.objectReferenceValue != null);
+                if (EditorGUI.EndChangeCheck())
+                {
+                    m_PostProcessData.objectReferenceValue = postProcessIncluded ? PostProcessData.GetDefaultPostProcessData() : null;
+                }
+                if (postProcessIncluded)
+                {
+                    EditorGUI.indentLevel++;
+                    EditorGUILayout.PropertyField(m_PostProcessData, Styles.postProcessLabel);
+                    EditorGUI.indentLevel--;
+                }
+
+                bool renderersUsePostProcessData = false;
+                for (int i = 0; i < m_RendererDataProp.arraySize; i++)
+                {
+                    var rendererData = m_RendererDataProp.GetArrayElementAtIndex(i).objectReferenceValue as ScriptableRendererData;
+
+                    var forwardRendererData = rendererData as ForwardRendererData;
+                    if (forwardRendererData != null && forwardRendererData.postProcessData != null)
+                    {
+                        renderersUsePostProcessData = true;
+                        break;
+                    }
+
+                    var renderer2DData = rendererData as UnityEngine.Experimental.Rendering.Universal.Renderer2DData;
+                    if (renderer2DData != null && renderer2DData.postProcessData != null)
+                    {
+                        renderersUsePostProcessData = true;
+                        break;
+                    }
+                }
+                if (!postProcessIncluded && renderersUsePostProcessData)
+                {
+                    EditorGUILayout.HelpBox("There are renderers that use custom Post Process Data for this reason post processing will not be disabled.", MessageType.Warning);
+                }
+
+                GUI.enabled = postProcessIncluded;
+
                 EditorGUILayout.PropertyField(m_ColorGradingMode, Styles.colorGradingMode);
                 if (!isHdrOn && m_ColorGradingMode.intValue == (int)ColorGradingMode.HighDynamicRange)
                     EditorGUILayout.HelpBox(Styles.colorGradingModeWarning, MessageType.Warning);
@@ -394,6 +437,8 @@ namespace UnityEditor.Rendering.Universal
                 m_ColorGradingLutSize.intValue = Mathf.Clamp(m_ColorGradingLutSize.intValue, UniversalRenderPipelineAsset.k_MinLutSize, UniversalRenderPipelineAsset.k_MaxLutSize);
                 if (isHdrOn && m_ColorGradingMode.intValue == (int)ColorGradingMode.HighDynamicRange && m_ColorGradingLutSize.intValue < 32)
                     EditorGUILayout.HelpBox(Styles.colorGradingLutSizeWarning, MessageType.Warning);
+
+                GUI.enabled = true;
 
                 EditorGUI.indentLevel--;
                 EditorGUILayout.Space();