+++ com.unity.render-pipelines.universal/Runtime/2D/Renderer2DDataAuthoring.cs
@@ -61,7 +61,10 @@ namespace UnityEngine.Experimental.Rendering.Universal
             }
 
             ResourceReloader.TryReloadAllNullIn(this, UniversalRenderPipelineAsset.packagePath);
-            ResourceReloader.TryReloadAllNullIn(m_PostProcessData, UniversalRenderPipelineAsset.packagePath);
+
+            // As now post process data is stored in Universal Render Pipeline, we can dereference non custom data.
+            if (m_PostProcessData == PostProcessData.GetDefaultPostProcessData())
+                m_PostProcessData = null;
         }
 
         private void Awake()